test_requirements.txt
flask==2.3.3  
werkzeug==2.3.7  
pytest==7.4.0  
pytest-mock==3.11.1  
email-validator==2.0.0  
cohere==4.34  
pymongo==4.5.0  
python-dotenv==1.0.0  
bcrypt==4.0.1  
pyjwt==2.8.0  
kafka-python==2.0.2 
